Renamed file tree from components to SRC and TestMap to LeafletMap
