+++ SRC/MAPS/LeafletMap.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import TestMap from './TestMap';
+import LeafletMap from './LeafletMap';
 
-describe('TestMap', () => {
+describe('LeafletMap', () => {
   const data = {
     position: [45.521, -122.664],
     zoom: 13,
@@ -10,7 +10,7 @@ describe('TestMap', () => {
     attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors',
   };
 
-  const wrapper = shallow(<TestMap
+  const wrapper = shallow(<LeafletMap
     position={data.position}
     zoom={data.zoom}
     url={data.url}
