+++ SRC/Slider/Slider.js
