+++ SRC/Slider/SliderTooltip.js
