+++ SRC/StoryCard/StoryFooter.css
