+++ SRC/index.js
@@ -6,7 +6,7 @@ import ChartData from './ChartData/ChartData';
 import DropdownMenu from './DropdownMenu/DropdownMenu';
 import Pie from './Pie/Pie';
 import Slider from './Slider/Slider';
-import TestMap from './TestMap/TestMap';
+import LeafletMap from './MAPS/LeafletMap';
 import Header from './Navigation/Header';
 import Nav from './Navigation/Nav';
 import NavRouterLink from './Navigation/NavRouterLink';
@@ -20,7 +20,7 @@ export {
   Header,
   Pie,
   Slider,
-  TestMap,
+  LeafletMap,
   StoryCard,
   Nav,
   NavRouterLink,
