+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "",
-    "test": "mocha --compilers js:babel-core/register --require ./mocha.conf.js --require jsdom-global/register --require ignore-styles --colors -w ./components/**/*.test.js",
+    "test": "mocha --compilers js:babel-core/register --require ./mocha.conf.js --require jsdom-global/register --require ignore-styles --colors -w ./SRC/**/*.test.js",
     "start": "start-storybook -p 6006",
     "lint": "eslint components stories --fix",
     "deploy-storybook": "storybook-to-ghpages",
