+++ stories/DropdownMenu.story.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@kadira/storybook';
-import { DropdownMenu } from '../components';
+import { DropdownMenu } from '../SRC';
 
 const displayName = DropdownMenu.displayName || 'DropdownMenu';
 const title = 'Simple usage';
